requirements.txt
# RAG Chatbot Requirements
# Compatible with Python 3.9+ (tested on Python 3.12)

# Core FastAPI and web server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Document processing
pdfplumber>=0.10.4
python-docx>=1.1.0
pandas>=2.1.4
openpyxl>=3.1.2

# OCR and image processing
pytesseract>=0.3.10
Pillow>=10.1.0

# Embeddings and vector store
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# LLM inference
transformers>=4.36.2
# NOTE: Install PyTorch separately with CUDA support:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
accelerate>=0.25.0

# Optional: Memory optimization for GPU
bitsandbytes>=0.41.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
numpy>=1.26.0  # Required for Python 3.12 compatibility
scikit-learn>=1.3.2
requests>=2.31.0

# Web interface
streamlit>=1.29.0
